Log4j1 to Log4j2
Consolidate ParameterizedLogging tests

Update Log4j 1.x and 2.x to SLF4J description messages to be more concise

Ensure log4j 2.x Logger fatal log level is parameterized

Fixup recipe description tense

Sidenote, update Logback recipe display name to say what it actually does

ParameterizedLogging should reference visited super method invocation

ParameterizedLogging should parameterize non-string single-argument objects using double-braces
