+++ src/main/java/org/openrewrite/java/logging/logback/Log4jLayoutToLogback.java
@@ -89,7 +89,7 @@ public class Log4jLayoutToLogback extends Recipe {
                             cd.getCoordinates().replaceExtendsClause()
                     );
 
-                    // should be covered by maybeAddImport, fixme
+                    // should be covered by maybeAddImport, todo
                     doAfterVisit(new AddImport<>("ch.qos.logback.core.LayoutBase", null, false));
                 }
